fix: target user when using api key
Before, it was impossible to perform a request related to user while using an api-key, because user-related requests were using the user's JWT Token passed in the headers in order to retrieve the user. Now the user's email/id can be passed in the request body if an api-key is used.